requirements.txt
Flask==2.0.1
cryptography==36.0.1
pycryptodome==3.12.0
rabbit-cipher==1.0.0